feat(journey-maps): revert unintentional deletion of map-route-service.ts
Refs: ROKAS-1204
