+++ src/journey-maps/angular/services/map/map-routes.service.ts
@@ -1,19 +1,19 @@
 import { Injectable } from '@angular/core';
-import { FeatureCollection, Point } from 'geojson';
-import { GeoJSONSource, Map as MaplibreMap } from 'maplibre-gl';
+import { Point } from 'geojson';
+import { Map as MaplibreMap } from 'maplibre-gl';
 
 import {
   SbbRouteMetaInformation,
   SbbSelectableFeatureCollection,
 } from '../../journey-maps.interfaces';
 import { SbbMarker } from '../../model/marker';
-import { SBB_ROKAS_ROUTE_SOURCE } from '../constants';
 
 import { SbbMapSelectionEvent, SBB_SELECTED_PROPERTY_NAME } from './events/map-selection-event';
 import {
   SBB_ROUTE_ID_PROPERTY_NAME,
   SBB_ROUTE_LINE_COLOR_PROPERTY_NAME,
 } from './events/route-utils';
+import { SbbMapRouteService } from './map-route-service';
 import { SBB_EMPTY_FEATURE_COLLECTION } from './map-service';
 
 export const SBB_ALL_ROUTE_LAYERS: string[] = [
@@ -27,6 +27,8 @@ export const SBB_ALL_ROUTE_LAYERS: string[] = [
 
 @Injectable({ providedIn: 'root' })
 export class SbbMapRoutesService {
+  constructor(private _mapRouteService: SbbMapRouteService) {}
+
   updateRoutes(
     map: MaplibreMap,
     mapSelectionEventService: SbbMapSelectionEvent,
@@ -48,7 +50,7 @@ export class SbbMapRoutesService {
         }
       }
     });
-    this.updateRoute(map, mapSelectionEventService, {
+    this._mapRouteService.updateRoute(map, mapSelectionEventService, {
       type: 'FeatureCollection',
       // With ES2019 we can replace this with routes.flatMap(({features}) => features)
       features: routes.reduce(
@@ -57,6 +59,7 @@ export class SbbMapRoutesService {
       ),
     });
   }
+
   getRouteMarkers(
     routes: SbbSelectableFeatureCollection[] | undefined,
     routesOptions: SbbRouteMetaInformation[] | undefined
@@ -81,20 +84,5 @@ export class SbbMapRoutesService {
         } as SbbMarker;
       })
       .filter((m) => !!m) as SbbMarker[];
-  }
-
-  updateRoute(
-    map: MaplibreMap,
-    mapSelectionEventService: SbbMapSelectionEvent,
-    routeFeatureCollection: FeatureCollection = SBB_EMPTY_FEATURE_COLLECTION
-  ): void {
-    const source = map.getSource(SBB_ROKAS_ROUTE_SOURCE) as GeoJSONSource;
-    source.setData(routeFeatureCollection);
-    map.removeFeatureState({ source: SBB_ROKAS_ROUTE_SOURCE });
-    if (routeFeatureCollection.features?.length) {
-      map.once('idle', () =>
-        mapSelectionEventService.initSelectedState(map, routeFeatureCollection.features, 'ROUTE')
-      );
-    }
   }
 }
